ci: Add external links for processes
Bug reports, Communications, and Feature requests can be referenced directly from GitHub issues and we can prevent issues being raised on the repos so these requests go to the right places.

Used the employee versions of the forms as the organisation is restricted and employees will be the ones in the repositories.